Fix Dancing Stage loading on videos
We had reports in the OutFox Discord server and on other sites
that the Dancing Stage was still being loaded, despite a video being
present. This was because of the loading process of the stages, which
has been tweaked here to prevent such case.

I would still suggest reworking how these stages are made, because
having Lua embedded to the stage to be loaded as global variables is
quite sketchy.